+++ src/components/list-modals/index.ts
@@ -1,6 +1,6 @@
 export { BaseEventListModal, type EventListAction, type EventListItem } from "./base-event-list-modal";
-export { DisabledRecurringEventsModal } from "./disabled-recurring-events-modal";
 export { FilteredEventsModal } from "./filtered-events-modal";
 export { GlobalSearchModal } from "./global-search-modal";
+export { RecurringEventsModal } from "./recurring-events-modal";
 export { SelectedEventsModal } from "./selected-events-modal";
 export { SkippedEventsModal } from "./skipped-events-modal";